requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
sqlalchemy==2.0.23
faker==20.1.0
pydantic==2.5.0
python-multipart==0.0.6
cryptography
python-dotenv==1.0.0
